attendence system/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
auto-py-to-exe==2.43.3
beautifulsoup4==4.12.3
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.90
Eel==0.16.0
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
future==1.0.0
gast==0.5.4
gdown==5.1.0
gevent==24.2.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.2
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
keras==3.3.2
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
mysql-connector-python==8.3.0
namex==0.0.8
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMySQL==1.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
retina-face==0.0.17
rich==13.7.1
setuptools==65.5.0
six==1.16.0
soupsieve==2.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-keras==2.16.0
tqdm==4.66.2
typing-extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
whichcraft==0.6.1
wrapt==1.16.0
zope.event==5.0
zope.interface==6.3
